[drm_atomic|fork]: Fix build when CONFIG_KPROFILES is disabled
Because the implementation of external functions in kprofiles is not
in accordance with the Linux kernel standards, we have 2 choices, namely
adding an dummy functions when !CONFIG_KPROFILES or adding the
CONFIG_KPROFILES macro to functions used outside. Since we're lazy to
add dummy functions, let's just add the CONFIG_KPROFILES macro to
fix build failure when CONFIG_KPROFILES is disabled.

Signed-off-by: Fiqri Ardyansyah <fiqri0927936@gmail.com>